Create stock_beta.py
Performed a linear regression and calculated correlations between 20 stocks that had a gain of at least 100% in 2021 with respect to the S&P500 ETF SPY. After plotting each linear regression, I made a heat map with the correlation values for the stocks and SPY. It showed what was expected. The stocks with huge gains were not correlated with the S&P500, but were correlated with each other. Most of the beta values were greater than 1, some greater than 2 which was expected as stocks performing this well are much more volatile than the S&P500 index.